test: 修改react-native-tts测试demo
